resurrect exclusive fullscreen mode to be able to support 10bit monitor output (limited to sRGB for now, HDR/Rec2020 will follow as soon as NVAPI supports it)
..same limitation for exclusive fullscreen mode as before: no other windows will show up, not even the exit dialogue  :/
..but at least ALT-TAB, etc is supported because of d3d9ex

git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2577 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pintable.cpp
@@ -898,7 +898,7 @@ void upscale(DWORD * const data, const unsigned int xres, const unsigned int yre
                     case 2: res = ((i <= 1) ? 0xdeadbeef : -2); break;
                     case 3: res = ((i == xres - 1) ? 0xdeadbeef : 1); break;
                     case 4: res = ((i >= xres - 2) ? 0xdeadbeef : 2); break;
-                    case 5: res = ((j == 0) ? 0xdeadbeef : -xres); break;
+                    case 5: res = ((j == 0) ? 0xdeadbeef : (unsigned int)(-(int)xres)); break;
                     case 6: res = ((j <= 1) ? 0xdeadbeef : -2 * xres); break;
                     case 7: res = ((j == yres - 1) ? 0xdeadbeef : xres); break;
                     case 8: res = ((j >= yres - 2) ? 0xdeadbeef : 2 * xres); break;
